+++ apps/DatabaseMigrations/Program.cs
@@ -65,7 +65,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using RadixDlt.NetworkGateway.PostgresIntegration;
 using RadixDlt.NetworkGateway.PostgresIntegration.IoC;
 using System.Threading.Tasks;