+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/ValidatorEmissionProcessor.cs
@@ -64,7 +64,6 @@
 
 using NpgsqlTypes;
 using RadixDlt.NetworkGateway.PostgresIntegration.Models;
-using RadixDlt.NetworkGateway.PostgresIntegration.Services;
 using RadixDlt.NetworkGateway.PostgresIntegration.Utils;
 using System.Collections.Generic;
 using System.Collections.Immutable;