+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Services/DapperWrapper.cs
@@ -73,7 +73,6 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Diagnostics;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;