+++ tests/RadixDlt.NetworkGateway.UnitTests/Abstractions/ValueBytesTests.cs
@@ -65,7 +65,6 @@
 using FluentAssertions;
 using RadixDlt.NetworkGateway.Abstractions;
 using RadixDlt.NetworkGateway.GatewayApiSdk.Model;
-using RadixDlt.NetworkGateway.PostgresIntegration;
 using RadixDlt.NetworkGateway.PostgresIntegration.Utils;
 using System;
 using System.Collections.Generic;