+++ tests/RadixDlt.NetworkGateway.UnitTests/PostgresIntegration/ScryptoSborUtilsTests.cs
@@ -63,7 +63,6 @@
  */
 
 using FluentAssertions;
-using RadixDlt.NetworkGateway.PostgresIntegration;
 using RadixDlt.NetworkGateway.PostgresIntegration.Utils;
 using System;
 using System.Collections.Generic;