+++ tests/RadixDlt.NetworkGateway.UnitTests/PostgresIntegration/TokenAmountExtensionsTests.cs
@@ -64,7 +64,6 @@
 
 using FluentAssertions;
 using RadixDlt.NetworkGateway.Abstractions.Numerics;
-using RadixDlt.NetworkGateway.PostgresIntegration;
 using RadixDlt.NetworkGateway.PostgresIntegration.Utils;
 using System.Linq;
 using System.Numerics;